move store cost computation function to util package and use it in server
- there's no reason this had to be in the frontend...
