+++ src/packages/next/components/store/other-items.tsx
@@ -18,7 +18,7 @@ import {
   Table,
 } from "antd";
 import { DisplayCost, describeItem } from "./site-license-cost";
-import { computeCost } from "./compute-cost";
+import { computeCost } from "@cocalc/util/licenses/store/compute-cost";
 import Loading from "components/share/loading";
 import { Icon } from "@cocalc/frontend/components/icon";
 import { search_split, search_match } from "@cocalc/util/misc";
