+++ src/packages/next/components/store/vouchers.tsx
@@ -33,7 +33,7 @@ import apiPost from "lib/api/post";
 import useAPI from "lib/hooks/api";
 import useIsMounted from "lib/hooks/mounted";
 import { useRouter } from "next/router";
-import { computeCost } from "./compute-cost";
+import { computeCost } from "@cocalc/util/licenses/store/compute-cost";
 import { useProfileWithReload } from "lib/hooks/profile";
 import { Paragraph } from "components/misc";
 import {
